fixing heath checks
