+++ .github/workflows/ci.yml
@@ -39,7 +39,7 @@ jobs:
       semVer: ${{ steps.gitversion.outputs.semVer }}
       preReleaseLabel: ${{ steps.gitversion.outputs.preReleaseLabel }}
       majorMinorPatch: ${{ steps.gitversion.outputs.majorMinorPatch }}
-      nuGetVersionV2: ${{ steps.gitversion.outputs.nuGetVersionV2 }}
+      nuGetVersionV2: ${{ steps.gitversion.outputs.MajorMinorPatch }}
 
     steps:
       - uses: actions/checkout@v4
@@ -50,12 +50,17 @@ jobs:
         with:
           dotnet-version: "8.0.x"
 
+      # - name: Install GitVersion
+      #   run: dotnet tool install --global GitVersion.Tool
+
       - name: Install GitVersion
-        run: dotnet tool install --global GitVersion.Tool
+        uses: gittools/actions/gitversion/setup@v2.0.1
+        with:
+          versionSpec: '6.x'
 
       - name: Determine Version
         id: gitversion
-        uses: gittools/actions/gitversion/execute@v0.10.2
+        uses: gittools/actions/gitversion/execute@v2.0.1
         with:
           useConfigFile: true
           updateAssemblyInfo: true
