+++ doc/dependency_decisions.yml
@@ -588,6 +588,7 @@
       - 8.0.2
       - 8.0.3
       - 8.0.6
+      - 8.0.7
     :when: 2022-10-14T23:37:16.793Z
     :who: mocsharp
     :why: MIT (https://github.com/dotnet/runtime/raw/main/LICENSE.TXT)
