+++ tests/Integration.Test/appsettings.json
@@ -27,8 +27,8 @@
     }
   },
   "ConnectionStrings": {
-    "Type": "mongodb",
-    "InformaticsGatewayDatabase": "mongodb://root:rootpassword@localhost:27017",
+    "Type": "sqlite",
+    "InformaticsGatewayDatabase": "Data Source=./mig.db",
     "DatabaseOptions": {
       "DatabaseName": "InformaticsGateway",
       "retries": {
